Create package lock.json
This is a package-lock.json file, which is a JSON file that contains information about the dependencies and their versions in a project. It is generated by npm (Node Package Manager) when installing packages.

Here's a breakdown of the file:

- name and version: The name and version of the project.
- lockfileVersion: The version of the lockfile format.
- requires: A boolean indicating whether the project requires a specific version of npm.
- packages: An object containing information about the packages in the project.
    - "": The root package (the project itself).
        - name and version: The name and version of the project.
        - dependencies: An object containing the dependencies of the project, with their versions.
        - devDependencies: An object containing the development dependencies of the project, with their versions.
- node_modules/@aashutoshrathi/word-wrap and node_modules/@ampproject/remapping: These are dependencies of the project, with their versions and other information.

This file ensures that the project's dependencies are installed with the correct versions, and it also provides information about the dependencies for other developers who may work on the project.